+++ src/test/java/com/ciphertool/genetics/algorithms/selection/modes/AlphaSelectorTest.java
@@ -28,6 +28,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyZeroInteractions;
 
 import java.lang.reflect.Field;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -64,16 +65,16 @@ public class AlphaSelectorTest {
 		List<Chromosome> individuals = new ArrayList<Chromosome>();
 
 		MockKeyedChromosome chromosome1 = new MockKeyedChromosome();
-		chromosome1.setFitness(2.0);
+		chromosome1.setFitness(BigDecimal.valueOf(2.0));
 		individuals.add(chromosome1);
 
-		Double bestFitness = 3.0;
+		BigDecimal bestFitness = BigDecimal.valueOf(3.0);
 		MockKeyedChromosome chromosome2 = new MockKeyedChromosome();
 		chromosome2.setFitness(bestFitness);
 		individuals.add(chromosome2);
 
 		MockKeyedChromosome chromosome3 = new MockKeyedChromosome();
-		chromosome3.setFitness(1.0);
+		chromosome3.setFitness(BigDecimal.valueOf(1.0));
 		individuals.add(chromosome3);
 
 		int selectedIndex = alphaSelector.getNextIndex(individuals, null);
@@ -85,7 +86,7 @@ public class AlphaSelectorTest {
 
 	@Test
 	public void testGetNextIndexWithNullPopulation() {
-		int selectedIndex = alphaSelector.getNextIndex(null, 6.0);
+		int selectedIndex = alphaSelector.getNextIndex(null, BigDecimal.valueOf(6.0));
 
 		assertEquals(-1, selectedIndex);
 		verify(logMock, times(1)).warn(anyString());
@@ -93,7 +94,7 @@ public class AlphaSelectorTest {
 
 	@Test
 	public void testGetNextIndexWithEmptyPopulation() {
-		int selectedIndex = alphaSelector.getNextIndex(new ArrayList<Chromosome>(), 6.0);
+		int selectedIndex = alphaSelector.getNextIndex(new ArrayList<Chromosome>(), BigDecimal.valueOf(6.0));
 
 		assertEquals(-1, selectedIndex);
 		verify(logMock, times(1)).warn(anyString());
